+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/data/datastore/DataStoreSettingsRepository.kt
@@ -36,6 +36,7 @@ class DataStoreSettingsRepository(private val dataStoreManager: DataStoreManager
 	private val welcomeScreenCompleted = booleanPreferencesKey("WELCOME_SCREEN_COMPLETE")
 	private val quicEnabled = booleanPreferencesKey("QUIC_ENABLED")
 	private val isStreamingServerBannerDisplayed = booleanPreferencesKey("STREAMING_SERVER_DISPLAYED")
+	private val isPerAppSecurityBannerDisplayed = booleanPreferencesKey("DEFAULT_PER_APP_SECURITY_BANNER_DISPLAYED")
 
 	override suspend fun setEntryPoint(entry: EntryPoint) {
 		dataStoreManager.saveToDataStore(entryPoint, entry.asString())
@@ -195,6 +196,14 @@ class DataStoreSettingsRepository(private val dataStoreManager: DataStoreManager
 		dataStoreManager.saveToDataStore(this.isStreamingServerBannerDisplayed, displayed)
 	}
 
+	override suspend fun getIsPerAppSecurityBannerDisplayed(): Boolean {
+		return dataStoreManager.getFromStore(isPerAppSecurityBannerDisplayed) ?: Settings.DEFAULT_STREAMING_SERVER_BANNER_DISPLAYED
+	}
+
+	override suspend fun setIsPerAppSecurityBannerDisplayed(displayed: Boolean) {
+		dataStoreManager.saveToDataStore(this.isPerAppSecurityBannerDisplayed, displayed)
+	}
+
 	override val settingsFlow: Flow<Settings> =
 		dataStoreManager.preferencesFlow.map { prefs ->
 			prefs?.let { pref ->
@@ -223,6 +232,7 @@ class DataStoreSettingsRepository(private val dataStoreManager: DataStoreManager
 						welcomeScreenCompleted = pref[welcomeScreenCompleted] ?: Settings.FLAG_WELCOME_SCREEN_COMPLETED,
 						quicEnabled = pref[quicEnabled] ?: Settings.DEFAULT_QUIC_ENABLED,
 						isStreamingServerBannerDisplayed = pref[isStreamingServerBannerDisplayed] ?: Settings.DEFAULT_STREAMING_SERVER_BANNER_DISPLAYED,
+						isPerAppSecurityBannerDisplayed = pref[isPerAppSecurityBannerDisplayed] ?: Settings.DEFAULT_PER_APP_SECURITY_BANNER_DISPLAYED,
 					)
 				} catch (e: IllegalArgumentException) {
 					Timber.e(e)