+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/common/animations/PulsingDots.kt
@@ -23,14 +23,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 
 @Composable
-fun PulsingDotsWave(
-	modifier: Modifier = Modifier,
-	dotCount: Int = 4,
-	dotSize: Dp = 8.dp,
-	dotColor: Color = Color(0xFF00FF66),
-	spaceBetween: Dp = 2.dp,
-	pulseDuration: Int = 1000,
-) {
+fun PulsingDotsWave(modifier: Modifier = Modifier, dotCount: Int = 4, dotSize: Dp = 8.dp, dotColor: Color = Color(0xFF00FF66), spaceBetween: Dp = 2.dp, pulseDuration: Int = 1000) {
 	val infiniteTransition = rememberInfiniteTransition(label = "waveTransition")
 	val progress by infiniteTransition.animateFloat(
 		initialValue = 0f,