+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/details/components/DetailsSectionPrivacy.kt
@@ -31,12 +31,7 @@ import net.nymtech.nymvpn.ui.theme.Typography
 import nym_vpn_lib_types.AsnKind
 
 @Composable
-fun DetailsSectionPrivacy(
-	asnKind: AsnKind?,
-	isQuicFeatureFlagEnabled: Boolean,
-	isQuicSupportedByGateway: Boolean,
-	onEnableQuicProtocolClick: () -> Unit,
-) {
+fun DetailsSectionPrivacy(asnKind: AsnKind?, isQuicFeatureFlagEnabled: Boolean, isQuicSupportedByGateway: Boolean, onEnableQuicProtocolClick: () -> Unit) {
 	val items = buildList<Pair<String, @Composable () -> Unit>> {
 		add(
 			stringResource(R.string.details_advanced_privacy) to {