+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/ui/screens/main/MainScreen.kt
@@ -113,6 +113,7 @@ fun MainScreen(appUiState: AppUiState, autoStart: Boolean, viewModel: MainViewMo
 	var showNetworkStatsDialog by remember { mutableStateOf(false) }
 	val isAppInForeground by viewModel.isAppInForeground.collectAsState()
 	var showBanner by remember { mutableStateOf(false) }
+	var showPerAppSecurityBanner by remember { mutableStateOf(false) }
 
 	with(appUiState.managerState) {
 		LaunchedEffect(tunnelState, connectionData?.connectedAt) {
@@ -201,6 +202,10 @@ fun MainScreen(appUiState: AppUiState, autoStart: Boolean, viewModel: MainViewMo
 		viewModel.onStreamingServerBannerDisplayed()
 		showBanner = false
 	}
+	fun dismissPerAppSecurityBanner() {
+		viewModel.onPerAppSecurityBannerDisplayed()
+		showPerAppSecurityBanner = false
+	}
 
 	fun onEntryClick() {
 		when (uiState.connectionState) {
@@ -221,7 +226,9 @@ fun MainScreen(appUiState: AppUiState, autoStart: Boolean, viewModel: MainViewMo
 	}
 
 	LaunchedEffect(Unit) {
-		if (!appUiState.settings.isStreamingServerBannerDisplayed) {
+		if (!appUiState.settings.isPerAppSecurityBannerDisplayed) {
+			showPerAppSecurityBanner = true
+		} else if (!appUiState.settings.isStreamingServerBannerDisplayed) {
 			showBanner = true
 		}
 		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
@@ -333,6 +340,26 @@ fun MainScreen(appUiState: AppUiState, autoStart: Boolean, viewModel: MainViewMo
 					icon = Icons.Outlined.Close,
 					onClicked = {
 						dismissStreamingBanner()
+					},
+				),
+			),
+			modifier = Modifier.padding(16.dp),
+		)
+		InfoBanner(
+			showBanner = showPerAppSecurityBanner,
+			config = BannerConfig(
+				message = stringResource(R.string.split_tunneling_per_app_security_banner_title),
+				action = BannerAction(
+					title = stringResource(R.string.split_tunneling_per_app_security_banner_action),
+					onClicked = {
+						dismissPerAppSecurityBanner()
+						navController.goFromRoot(Route.SplitTunneling)
+					},
+				),
+				icon = BannerIcon(
+					icon = Icons.Outlined.Close,
+					onClicked = {
+						dismissPerAppSecurityBanner()
 					},
 				),
 			),