+++ nym-vpn-android/core/src/main/java/net/nymtech/vpn/backend/Backend.kt
@@ -32,7 +32,7 @@ interface Backend {
 
 	suspend fun updateAccountState()
 
-	suspend fun start(tunnel: Tunnel, userAgent: UserAgent, enableBridges: Boolean)
+	suspend fun start(tunnel: Tunnel, userAgent: UserAgent, enableBridges: Boolean, restrictedAppsPackages: List<String> = emptyList())
 
 	suspend fun stop()